+++ soda-tests/tests/components/test_cli_contract_handlers.py
@@ -1,13 +1,6 @@
-from soda_core.common.exceptions import SodaCloudException
-from unittest.mock import MagicMock, patch
-
 import pytest
 
 from soda_core.cli.exit_codes import ExitCode
-from soda_core.contracts.api.verify_api import (
-    ContractVerificationSession,
-    all_none_or_empty,
-)
 from soda_core.cli.handlers.contract import (
     handle_publish_contract,
     handle_test_contract,
@@ -19,14 +12,8 @@ from soda_core.contracts.contract_publication import (
     ContractPublicationResult,
     ContractPublicationResultList,
 )
-
-
-@pytest.mark.parametrize(
-    "array, expected", [(["a", "b"], False), (["a", None], False), ([None, None], False), ([], True), (None, True)]
-)
-def test_all_none_or_empty(array, expected):
-    is_all_none_or_empty = all_none_or_empty(array)
-    assert is_all_none_or_empty == expected
+from soda_core.contracts.contract_verification import ContractVerificationSession
+from unittest.mock import MagicMock, patch
 
 
 @pytest.mark.parametrize(
@@ -68,148 +55,6 @@ def test_handle_verify_contract_exit_codes(
     )
 
     assert exit_code == expected_exit_code
-
-
-def test_handle_verify_contract_returns_exit_code_3_when_using_dataset_names_without_cloud_configuration(caplog):
-    exit_code = handle_verify_contract(
-        contract_file_paths=None,
-        dataset_identifiers=["some_dataset"],
-        data_source_file_path="ds.yaml",
-        soda_cloud_file_path=None,
-        variables={},
-        publish=False,
-        use_agent=False,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-    )
-    assert (
-        "A Soda Cloud configuration file is required to use the -d/--dataset argument."
-        "Please provide the '--soda-cloud' argument with a valid configuration file path."
-    ) in caplog.messages
-    assert exit_code == ExitCode.LOG_ERRORS
-
-
-def test_handle_verify_contract_returns_exit_code_3_when_using_publish_without_cloud_configuration(caplog):
-    exit_code = handle_verify_contract(
-        contract_file_paths=None,
-        dataset_identifiers=["some_dataset"],
-        data_source_file_path="ds.yaml",
-        soda_cloud_file_path=None,
-        variables={},
-        publish=True,
-        use_agent=False,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-    )
-
-    assert (
-        "A Soda Cloud configuration file is required to use the -p/--publish argument. "
-        "Please provide the '--soda-cloud' argument with a valid configuration file path."
-    ) in caplog.messages
-    assert exit_code == ExitCode.LOG_ERRORS
-
-
-@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
-def test_handle_verify_contract_returns_exit_code_3_when_no_contract_file_paths_or_dataset_identifiers(
-    mock_cloud_client, caplog
-):
-    exit_code = handle_verify_contract(
-        contract_file_paths=None,
-        dataset_identifiers=None,
-        data_source_file_path="ds.yaml",
-        soda_cloud_file_path="sc.yaml",
-        variables={},
-        publish=True,
-        use_agent=False,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-    )
-
-    assert "At least one of -c/--contract or -d/--dataset arguments is required." in caplog.messages
-    assert exit_code == ExitCode.LOG_ERRORS
-
-
-@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
-def test_handle_verify_contract_returns_exit_code_3_when_no_data_source_configuration_or_dataset_identifiers(
-    mock_cloud_client, caplog
-):
-    exit_code = handle_verify_contract(
-        contract_file_paths=["contract.yaml"],
-        dataset_identifiers=None,
-        data_source_file_path=None,
-        soda_cloud_file_path="sc.yaml",
-        variables={},
-        publish=True,
-        use_agent=False,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-    )
-
-    assert "At least one of -ds/--data-source or -d/--dataset value is required." in caplog.messages
-    assert exit_code == ExitCode.LOG_ERRORS
-
-
-@pytest.mark.skip(reason="Needs mocking of Contract verification.")
-def test_handle_verify_contract_returns_exit_code_0_when_no_data_source_configuration_or_dataset_identifiers_and_remote(
-    caplog,
-):
-    exit_code = handle_verify_contract(
-        contract_file_paths=["contract.yaml"],
-        dataset_identifiers=None,
-        data_source_file_path=None,
-        variables={},
-        publish=True,
-        use_agent=True,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-        soda_cloud_client=MagicMock(),
-    )
-
-    assert "At least one of -ds/--data-source or -d/--dataset value is required." not in caplog.messages
-    assert exit_code == ExitCode.OK
-
-
-@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
-def test_handle_verify_contract_skips_contract_when_contract_fetching_from_cloud_returns_errors(
-    mock_cloud_client, caplog
-):
-    mock_cloud_client.return_value.fetch_contract_for_dataset.side_effect = SodaCloudException("woopsie")
-
-    _ = handle_verify_contract(
-        contract_file_paths=None,
-        dataset_identifiers=["my/super/awesome/identifier"],
-        data_source_file_path="ds.yaml",
-        soda_cloud_file_path="sc.yaml",
-        variables={},
-        publish=True,
-        use_agent=False,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-    )
-
-    assert (
-        "Could not fetch contract for dataset 'my/super/awesome/identifier': skipping verification" in caplog.messages
-    )
-
-
-@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
-def test_handle_verify_contract_returns_exit_code_0_when_no_valid_remote_contracts_left(mock_cloud_client, caplog):
-    mock_cloud_client.return_value.fetch_contract_for_dataset.side_effect = SodaCloudException("woopsie")
-
-    exit_code = handle_verify_contract(
-        contract_file_paths=None,
-        dataset_identifiers=["my/super/awesome/identifier"],
-        data_source_file_path="ds.yaml",
-        soda_cloud_file_path="sc.yaml",
-        variables={},
-        publish=True,
-        use_agent=False,
-        verbose=False,
-        blocking_timeout_in_minutes=10,
-    )
-
-    assert "No contracts given. Exiting." in caplog.messages
-    assert exit_code == ExitCode.OK
 
 
 @pytest.mark.parametrize(
