+++ soda-tests/tests/components/test_contract_verification_api.py
@@ -1,11 +1,21 @@
 import pytest
-from soda_core.common.exceptions import InvalidDatasetQualifiedNameException, YamlParserException
+
+from soda_core.common.exceptions import (
+    YamlParserException,
+    InvalidArgumentException,
+    SodaCloudException,
+)
+from soda_core.common.soda_cloud import SodaCloud
 from soda_core.common.yaml import ContractYamlSource
-from soda_core.contracts.contract_verification import (
+from soda_core.contracts.api.verify_api import (
     ContractVerificationSession,
+    all_none_or_empty, verify_contracts,
+)
+from soda_core.contracts.contract_verification import (
     ContractVerificationSessionResult,
     SodaException,
 )
+from unittest.mock import patch, MagicMock
 
 
 def test_contract_verification_file_api():
@@ -52,24 +62,155 @@ def test_contract_provided_and_configured():
     assert "Data source 'abc' not found" in contract_verification_session_result.get_errors_str()
 
 
-# TODO @Niels: To be evaluated if still needed refactored after rework
-# @pytest.mark.parametrize(
-#     "logs, contract_failed, has_errors, has_failures",
-#     [
-#         (Logs([Log(level=CRITICAL, message="critical")]), False, True, False),
-#         (Logs([Log(level=ERROR, message="error")]), False, True, False),
-#         (Logs([Log(level=ERROR, message="error"), Log(level=CRITICAL, message="critical")]), False, True, False),
-#         (Logs([Log(level=WARN, message="warn"), Log(level=INFO, message="info")]), False, False, False),
-#         (Logs([Log(level=WARN, message="warn"), Log(level=INFO, message="info")]), True, False, True),
-#         (Logs([Log(level=ERROR, message="error"), Log(level=CRITICAL, message="critical")]), True, True, True),
-#     ],
-# )
-# def test_contract_verification_log_levels(logs, contract_failed, has_errors, has_failures):
-#     mock_contract_result = Mock(spec=ContractResult)
-#     mock_contract_result.logs = logs
-#     mock_contract_result.failed.return_value = contract_failed
-#
-#     result = ContractVerificationResult(logs=logs, contract_results=[mock_contract_result])
-#
-#     assert result.has_errors() is has_errors
-#     assert result.has_failures() is has_failures
+@pytest.mark.parametrize(
+    "array, expected", [(["a", "b"], False), (["a", None], False), ([None, None], False), ([], True), (None, True)]
+)
+def test_all_none_or_empty(array, expected):
+    is_all_none_or_empty = all_none_or_empty(array)
+    assert is_all_none_or_empty == expected
+
+
+def test_handle_verify_contract_raises_exception_when_using_dataset_names_without_cloud_configuration():
+    with pytest.raises(
+        InvalidArgumentException,
+        match="A Soda Cloud configuration file is required to use the -d/--dataset argument."
+        "Please provide the '--soda-cloud' argument with a valid configuration file path."
+    ):
+        _ = verify_contracts(
+            contract_file_paths=None,
+            dataset_identifiers=["some_dataset"],
+            data_source_file_path="ds.yaml",
+            soda_cloud_file_path=None,
+            variables={},
+            publish=False,
+            use_agent=False,
+            verbose=False,
+            blocking_timeout_in_minutes=10,
+        )
+
+def test_handle_verify_contract_returns_exit_code_3_when_using_publish_without_cloud_configuration():
+    with pytest.raises(
+        InvalidArgumentException,
+        match="A Soda Cloud configuration file is required to use the -p/--publish argument. "
+              "Please provide the '--soda-cloud' argument with a valid configuration file path."
+    ):
+        _ = verify_contracts(
+            contract_file_paths=None,
+            dataset_identifiers=["some_dataset"],
+            data_source_file_path="ds.yaml",
+            soda_cloud_file_path=None,
+            variables={},
+            publish=True,
+            use_agent=False,
+            verbose=False,
+            blocking_timeout_in_minutes=10,
+        )
+
+
+@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
+def test_handle_verify_contract_returns_exit_code_3_when_no_contract_file_paths_or_dataset_identifiers(
+    mock_cloud_client
+):
+    with pytest.raises(
+        InvalidArgumentException,
+        match="At least one of -c/--contract or -d/--dataset arguments is required."
+    ):
+        _ = verify_contracts(
+            contract_file_paths=None,
+            dataset_identifiers=None,
+            data_source_file_path="ds.yaml",
+            soda_cloud_file_path="sc.yaml",
+            variables={},
+            publish=True,
+            use_agent=False,
+            verbose=False,
+            blocking_timeout_in_minutes=10,
+        )
+
+
+@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
+def test_handle_verify_contract_returns_exit_code_3_when_no_data_source_configuration_or_dataset_identifiers(
+    mock_cloud_client
+):
+    with pytest.raises(
+        InvalidArgumentException,
+        match="At least one of -ds/--data-source or -d/--dataset value is required."
+    ):
+        _ = verify_contracts(
+            contract_file_paths=["contract.yaml"],
+            dataset_identifiers=None,
+            data_source_file_path=None,
+            soda_cloud_file_path="sc.yaml",
+            variables={},
+            publish=True,
+            use_agent=False,
+            verbose=False,
+            blocking_timeout_in_minutes=10,
+        )
+
+
+@pytest.mark.skip(reason="Needs mocking of Contract verification.")
+@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
+def test_handle_verify_contract_returns_exit_code_0_when_no_data_source_configuration_or_dataset_identifiers_and_remote(
+    mock_cloud_client
+):
+    mock_cloud_client.return_value = MagicMock(spec=SodaCloud)
+
+    try:
+        _ = verify_contracts(
+            contract_file_paths=["contract.yaml"],
+            dataset_identifiers=None,
+            data_source_file_path=None,
+            soda_cloud_file_path="sc.yaml",
+            variables={},
+            publish=True,
+            use_agent=True,
+            verbose=False,
+            blocking_timeout_in_minutes=10,
+        )
+    except Exception as exc:
+        pytest.fail(f"An unexpected exception was raised: {exc}")
+
+
+
+
+@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
+def test_handle_verify_contract_skips_contract_when_contract_fetching_from_cloud_returns_errors(
+    mock_cloud_client, caplog
+):
+    mock_cloud_client.return_value.fetch_contract_for_dataset.side_effect = SodaCloudException("woopsie")
+
+    _ = verify_contracts(
+        contract_file_paths=None,
+        dataset_identifiers=["my/super/awesome/identifier"],
+        data_source_file_path="ds.yaml",
+        soda_cloud_file_path="sc.yaml",
+        variables={},
+        publish=True,
+        use_agent=False,
+        verbose=False,
+        blocking_timeout_in_minutes=10,
+    )
+
+    assert (
+        "Could not fetch contract for dataset 'my/super/awesome/identifier': skipping verification" in caplog.messages
+    )
+
+
+@patch("soda_core.contracts.api.verify_api.SodaCloud.from_config")
+def test_handle_verify_contract_returns_exit_code_0_when_no_valid_remote_contracts_left(mock_cloud_client, caplog):
+    mock_cloud_client.return_value.fetch_contract_for_dataset.side_effect = SodaCloudException("woopsie")
+
+    _ = verify_contracts(
+        contract_file_paths=None,
+        dataset_identifiers=["my/super/awesome/identifier"],
+        data_source_file_path="ds.yaml",
+        soda_cloud_file_path="sc.yaml",
+        variables={},
+        publish=True,
+        use_agent=False,
+        verbose=False,
+        blocking_timeout_in_minutes=10,
+    )
+
+    assert "No contracts given. Exiting." in caplog.messages
